refactor: Reorganizar el context de Year
