+++ src/admin/contexts/yearContext/YearContext.types.ts
@@ -1,4 +1,4 @@
-import type { CreateYearPayload } from "../services/Year/YearService";
+import type { CreateYearPayload } from "../../services/year/YearService";
 
 export interface YearContextType {
   loading: boolean;
